# docs/swagger.yaml
basePath: /
definitions:
  books.AvailabilityResponse:
    properties:
      price_in_kopeks:
        type: integer
      sku_uuid:
        type: string
      stock_count:
        type: integer
      store_name:
        type: string
      store_uuid:
        type: string
    type: object
  books.BookResponse:
    properties:
      author:
        type: string
      description:
        type: string
      id:
        type: integer
      isbn:
        type: string
      page_count:
        type: integer
      publication_year:
        type: integer
      title:
        type: string
    type: object
  books.CreateBookRequest:
    properties:
      author:
        type: string
      description:
        type: string
      isbn:
        type: string
      page_count:
        type: integer
      publication_year:
        type: integer
      title:
        type: string
    required:
      - author
      - isbn
      - title
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: { }
  description: Это REST API для сервиса сети книжных магазинов.
  title: Bookstores API
  version: "1.0"
paths:
  /books:
    get:
      description: Возвращает список всех книг в глобальном каталоге.
      produces:
        - application/json
      responses:
        "200":
          description: Список книг
          schema:
            items:
              $ref: '#/definitions/books.BookResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получить глобальный список книг
      tags:
        - books
    post:
      consumes:
        - application/json
      description: Создаёт новую книгу в глобальном каталоге. Если книга с таким ISBN
        уже существует, обновляет ее данные.
      parameters:
        - description: Данные для создания книги
          in: body
          name: input
          required: true
          schema:
            $ref: '#/definitions/books.CreateBookRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Инфо об книге
          schema:
            $ref: '#/definitions/books.BookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Создать новую книгу
      tags:
        - books
  /books/{bookID}:
    get:
      description: Возвращает информацию о книге по её ID.
      parameters:
        - description: ID книги
          in: path
          name: bookID
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Инфо о книге
          schema:
            $ref: '#/definitions/books.BookResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Книга отсутствует
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получить одну книгу
      tags:
        - books
  /books/{bookID}/availability:
    get:
      description: Показывает, в каких магазинах, по какой цене и в каком количестве
        доступна книга.
      parameters:
        - description: ID книги
          in: path
          name: bookID
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/books.AvailabilityResponse'
            type: array
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Книга отсутствует
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Доступность книги
      tags:
        - books
  /books/search:
    get:
      description: Ищет книги по части названия или имени автора.
      parameters:
        - description: Поисковый запрос
          in: query
          name: q
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Список найденных книг
          schema:
            items:
              $ref: '#/definitions/books.BookResponse'
            type: array
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Поиск книг
      tags:
        - books
...
swagger: "2.0"